refactor: reducers拆分
