+++ packages/builder/src/reducers/componentId.js
@@ -1,0 +1,15 @@
+export default (state = '', action) => {
+  let newState = state
+  const { data = {} } = action
+  const _componentId = data.componentId || ''
+  // eslint-disable-next-line
+  let { id, propsData = {}, comp = {}, key } = data
+
+  switch (action.type) {
+    case 'CHANGE_COMPONENT':
+      newState = _componentId
+      return newState
+    default:
+      return state
+  }
+}
