+++ packages/builder/src/reducers/preview.js
@@ -1,0 +1,12 @@
+export default (state = false, action) => {
+  let newState
+  const { data = {} } = action
+  const _preview = data.preview || false
+  switch (action.type) {
+    case 'CHANGE_PREVIEW':
+      newState = _preview
+      return newState
+    default:
+      return state
+  }
+}
